Change writer to connection
